Extracted SemanticModelRow and MetricRow into separate files and fixed lint
