+++ datahub-web-react/src/app/metrics/metricsTypes.ts
@@ -1,0 +1,12 @@
+import { GetSemanticModelMetricsQuery, GetSemanticModelsBrowseQuery } from '@graphql/metricsBrowse.generated';
+
+export type SemanticModel = NonNullable<
+    NonNullable<GetSemanticModelsBrowseQuery['getSemanticModels']>['semanticModels'][number]
+>;
+
+type SearchResultEntity = NonNullable<
+    NonNullable<
+        NonNullable<GetSemanticModelMetricsQuery['semanticModel']>['metrics']
+    >['searchResults'][number]['entity']
+>;
+export type MetricEntity = Extract<SearchResultEntity, { __typename?: 'Metric' }>;
